requirements.txt
rendercv[full]>=2.8
PyYAML>=6.0    # used by scripts/tailor.py for projection specs
